requirements.txt
telethon
pandas
sqlalchemy
psycopg2-binary
python-dotenv
ultralytics
dbt-core
dbt-postgres
fastapi
uvicorn
pydantic
dagster
dagster-webserver
